// frontend/src/types/enums.ts
export enum RouteLinks {
  landing = '/',
  login = '/login',
  ingredients = '/ingredients',
  results = '/results',
  recipe = '/recipe',
  favourites = '/favourites',
  history = '/history',
}

export enum dietReqs {
  glutenFree = "GF",
  dairyFree = "DF",
  vegetarian = "Vegetarian",
  vegan = "Vegan",
}